+++ VoiceCraft.Core/Audio/Effects/DirectionalEffect.cs
@@ -17,7 +17,7 @@ namespace VoiceCraft.Core.Audio.Effects
         {
         }
 
-        public virtual void Process(VoiceCraftEntity from, VoiceCraftEntity to, uint effectBitmask, Span<float> data,
+        public virtual void Process(VoiceCraftEntity from, VoiceCraftEntity to, ushort effectBitmask, Span<float> data,
             int count)
         {
             var bitmask = from.TalkBitmask & to.ListenBitmask & from.EffectBitmask & to.EffectBitmask;
