+++ VoiceCraft.Core/Audio/Effects/ProximityEffect.cs
@@ -12,7 +12,7 @@ namespace VoiceCraft.Core.Audio.Effects
         public int MaxRange { get; set; }
         public EffectType EffectType => EffectType.Proximity;
 
-        public virtual void Process(VoiceCraftEntity from, VoiceCraftEntity to, uint effectBitmask, Span<float> data,
+        public virtual void Process(VoiceCraftEntity from, VoiceCraftEntity to, ushort effectBitmask, Span<float> data,
             int count)
         {
             var bitmask = from.TalkBitmask & to.ListenBitmask & from.EffectBitmask & to.EffectBitmask;
@@ -48,7 +48,7 @@ namespace VoiceCraft.Core.Audio.Effects
             //Nothing to dispose.
         }
 
-        public bool Visibility(VoiceCraftEntity from, VoiceCraftEntity to, uint effectBitmask)
+        public bool Visibility(VoiceCraftEntity from, VoiceCraftEntity to, ushort effectBitmask)
         {
             var bitmask = from.TalkBitmask & to.ListenBitmask & from.EffectBitmask & to.EffectBitmask;
             if ((bitmask & effectBitmask) == 0) return true; //Proximity checking disabled.
