+++ VoiceCraft.Core/Interfaces/IAudioEffect.cs
@@ -8,7 +8,7 @@ namespace VoiceCraft.Core.Interfaces
     {
         EffectType EffectType { get; }
 
-        void Process(VoiceCraftEntity from, VoiceCraftEntity to, uint effectBitmask, Span<float> data, int count);
+        void Process(VoiceCraftEntity from, VoiceCraftEntity to, ushort effectBitmask, Span<float> data, int count);
 
         void Reset();
     }
